@@ -2,15 +2,15 @@ apply plugin: 'java'
 apply plugin: 'maven'
 apply plugin: 'signing'
 apply plugin: 'jacoco'
+apply plugin: 'org.junit.platform.gradle.plugin'
 
-
-sourceCompatibility = 1.7
+sourceCompatibility = 1.8
 compileJava.options.encoding = 'UTF-8'
 
 
 group = "com.github.zuinnote"
 archivesBaseName = "hadoopcryptoledger-hiveudf"
-version = "1.1.0"
+version = "1.1.1"
 
 
 
@@ -23,6 +23,25 @@ jacocoTestReport {
     }
 }
 
+
+
+// Integrate JaCoCo for jUnit5
+project.afterEvaluate {
+    def junitPlatformTestTask = project.tasks.getByName('junitPlatformTest')
+
+    // configure jacoco to analyze the junitPlatformTest task
+    jacoco {
+        applyTo junitPlatformTestTask
+    }
+
+    // create junit platform jacoco task
+    project.task(type: JacocoReport, "junitPlatformJacocoReport",
+            {
+                sourceDirectories = files("./src/main")
+                classDirectories = files("$buildDir/classes/main")
+                executionData junitPlatformTestTask
+            })
+}
 
 repositories {
     mavenCentral()
@@ -72,16 +91,18 @@ dependencies {
    // hadoop lib for driver
      provided("org.apache.hadoop:hadoop-client:2.7.0")
        // hadoop crypto ledger library
-   provided("com.github.zuinnote:hadoopcryptoledger-fileformat:1.1.0")
-   provided("com.github.zuinnote:hadoopcryptoledger-hiveserde:1.1.0")
+   provided("com.github.zuinnote:hadoopcryptoledger-fileformat:1.1.1")
+   provided("com.github.zuinnote:hadoopcryptoledger-hiveserde:1.1.1")
    // we need to mandatory add the BC library
      provided("org.bouncycastle:bcprov-ext-jdk15on:1.58")
     // hive serde API
    provided("org.apache.hive:hive-serde:1.2.0")
    provided("org.apache.hive:hive-exec:1.2.0")
    // log4j2
-   provided("org.apache.logging.log4j:log4j-api:2.4.1")
-    testCompile group: 'junit', name: 'junit', version: '4.12'
+   provided("org.apache.logging.log4j:log4j-api:2.9.1")
+   // test
+    testCompile group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: '5.0.2'
+	testRuntime group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: '5.0.2'
 }
 
 jar {
